refactor(site): render the sandbox chrome with React islands
Step 4 (PR A) of the site modernisation: React 19 arrives and the SANDBOX
page's chrome renders as islands mounted into the existing static shell.
The shell still paints from static HTML, CodeMirror still owns #editor, and
the IDE keeps its imperative components until PR B converts it.

The panels' imperative DOM juggling becomes declarative rendering:

- status-bar-store.ts  — the status bar's logic (problems, the rAF-coalesced
  500-line output log, executions), implementing the same StatusBar handle
  its producers already call.
- components/StatusBar.tsx — its view; same markup, same class names, all
  three lists kept mounted and toggled with `display` as before.
- runtime-target-core.ts — the external-runtime client + state machine,
  lifted out of runtime-target.ts unchanged; DOM writes became a snapshot.
- components/RuntimeTargetPanel.tsx — its React view (the sandbox);
  runtime-target.ts keeps the identical imperative view for the IDE.
- components/SandboxControls.tsx — the picker, reset, the device panel and
  the status pill.
- store.ts — the ~20-line external store the islands subscribe to.

Seams are preserved verbatim: window.__sandbox/__lang keep their shapes and
every e2e suite passes unmodified. The endpoint input stays UNCONTROLLED
with native listeners (the e2e assigns .value and dispatches an untrusted
`change`, which React's synthetic onChange would not see), the runtime-link
controller is created once outside React so its queued pushes and /state
poll chain survive re-renders, and `details.open` stays an imperative write.
