+++ package.json
@@ -53,10 +53,14 @@
     "@lezer/highlight": "^1.2.3",
     "@playwright/test": "^1.61.0",
     "@resvg/resvg-js": "^2.6.2",
+    "@types/react": "^19.2.17",
+    "@types/react-dom": "^19.2.3",
     "@vscode/test-electron": "^2.4.1",
     "codemirror": "^6.0.2",
     "esbuild": "^0.28.1",
     "png-to-ico": "^3.0.2",
+    "react": "^19.2.8",
+    "react-dom": "^19.2.8",
     "typescript": "^5.9.3"
   }
 }
